#374 Move all column logic together
