+++ api/src/main/java/com/ford/labs/retroquest/column/ColumnService.java
@@ -1,7 +1,10 @@
 package com.ford.labs.retroquest.column;
 
-import com.ford.labs.retroquest.columntitle.ColumnTitleRepository;
-import com.ford.labs.retroquest.columntitle.ColumnTitleService;
+import com.ford.labs.retroquest.exception.ColumnTitleNotFoundException;
+import com.ford.labs.retroquest.websocket.WebsocketColumnTitleEvent;
+import com.ford.labs.retroquest.websocket.WebsocketEventType;
+import com.ford.labs.retroquest.websocket.WebsocketService;
+import io.micrometer.core.instrument.MeterRegistry;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,9 +15,13 @@ import static java.util.stream.Collectors.toList;
 public class ColumnService {
 
     private final ColumnTitleRepository columnTitleRepository;
+    private final MeterRegistry meterRegistry;
+    private final WebsocketService websocketService;
 
-    public ColumnService(ColumnTitleRepository columnTitleRepository) {
+    public ColumnService(ColumnTitleRepository columnTitleRepository, MeterRegistry meterRegistry, WebsocketService websocketService) {
         this.columnTitleRepository = columnTitleRepository;
+        this.meterRegistry = meterRegistry;
+        this.websocketService = websocketService;
     }
 
     public List<Column> getColumns(String teamId) {
@@ -27,5 +34,18 @@ public class ColumnService {
                     return 0;
                 })
                 .collect(toList());
+    }
+
+    public ColumnTitle editColumnTitleName(Long columnId, String newColumnName, String teamId) {
+        var existingColumnTitle = columnTitleRepository.findById(columnId).orElseThrow(ColumnTitleNotFoundException::new);
+        existingColumnTitle.setTitle(newColumnName);
+
+        ColumnTitle newColumnTitle = columnTitleRepository.save(existingColumnTitle);
+
+        websocketService.publishEvent(new WebsocketColumnTitleEvent(teamId, WebsocketEventType.UPDATE, newColumnTitle));
+
+        meterRegistry.counter("retroquest.columns.changed.count").increment();
+
+        return newColumnTitle;
     }
 }
