+++ api/src/main/java/com/ford/labs/retroquest/column/UpdateColumnTitleRequest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.ford.labs.retroquest.columntitle;
+package com.ford.labs.retroquest.column;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Value;
