+++ api/src/main/java/com/ford/labs/retroquest/v2/columncombiner/columncombiner/ColumnCombinerService.java
@@ -18,8 +18,8 @@
 package com.ford.labs.retroquest.v2.columncombiner.columncombiner;
 
 import com.ford.labs.retroquest.actionitem.ActionItemRepository;
-import com.ford.labs.retroquest.columntitle.ColumnTitle;
-import com.ford.labs.retroquest.columntitle.ColumnTitleRepository;
+import com.ford.labs.retroquest.column.ColumnTitle;
+import com.ford.labs.retroquest.column.ColumnTitleRepository;
 import com.ford.labs.retroquest.thought.Thought;
 import com.ford.labs.retroquest.thought.ThoughtRepository;
 import org.springframework.stereotype.Service;
