+++ api/src/test/java/com/ford/labs/retroquest/deprecated_tests/TeamServiceTest.java
@@ -17,8 +17,8 @@
 
 package com.ford.labs.retroquest.deprecated_tests;
 
-import com.ford.labs.retroquest.columntitle.ColumnTitle;
-import com.ford.labs.retroquest.columntitle.ColumnTitleRepository;
+import com.ford.labs.retroquest.column.ColumnTitle;
+import com.ford.labs.retroquest.column.ColumnTitleRepository;
 import com.ford.labs.retroquest.exception.BoardDoesNotExistException;
 import com.ford.labs.retroquest.exception.PasswordInvalidException;
 import com.ford.labs.retroquest.team.*;
