+++ api/src/test/java/com/ford/labs/retroquest/v2/columncombiner/ColumnCombinerServiceTest.java
@@ -19,8 +19,8 @@ package com.ford.labs.retroquest.v2.columncombiner;
 
 import com.ford.labs.retroquest.actionitem.ActionItem;
 import com.ford.labs.retroquest.actionitem.ActionItemRepository;
-import com.ford.labs.retroquest.columntitle.ColumnTitle;
-import com.ford.labs.retroquest.columntitle.ColumnTitleRepository;
+import com.ford.labs.retroquest.column.ColumnTitle;
+import com.ford.labs.retroquest.column.ColumnTitleRepository;
 import com.ford.labs.retroquest.thought.Thought;
 import com.ford.labs.retroquest.thought.ThoughtRepository;
 import com.ford.labs.retroquest.v2.columncombiner.columncombiner.ColumnCombinerResponse;
